+++ NEWS.md
@@ -4,6 +4,8 @@
 
 ## Updates of Existing Functions
 
+- `assert_function()` now accepts all function arguments if ellipsis, `...` is in the function formals (#339)
+
 ## Breaking Changes
 
 - The following functions have been deprecated from previous admiral versions using the next phase of the deprecation process